requirements.txt
# UI and interaction
streamlit

# LangChain core components
langchain
langgraph
langchain-google-genai

# Vector store client
weaviate-client<4.0.0

# Tokenizer for Gemini models
tiktoken

# Load .env config
python-dotenv